deallocate/reallocate global arrays upon multiple calls to MECHANIC and "K" routines
git-svn-id: svn://dasher.wustl.edu/tinker@48 9dbd3ec4-0870-4ad8-bd6a-181c4f7deb04

+++ source/sktstuff.f
@@ -253,10 +253,10 @@ c
       implicit none
       integer i
       integer flag
-      real*8, allocatable :: b1(:)
-      real*8, allocatable :: b2(:)
-      real*8, allocatable :: b3(:)
-      real*8, allocatable :: b4(:)
+      integer, allocatable :: b1(:)
+      integer, allocatable :: b2(:)
+      integer, allocatable :: b3(:)
+      integer, allocatable :: b4(:)
 c
 c
 c     set initialization flag and test for socket usage
